+++ src/main/kotlin/com/github/njuro/postgresvaultgithub/VaultAuth.kt
@@ -2,6 +2,7 @@ package com.github.njuro.postgresvaultgithub
 
 import com.fasterxml.jackson.core.JsonProcessingException
 import com.intellij.credentialStore.Credentials
+import com.intellij.credentialStore.OneTimeString
 import com.intellij.database.access.DatabaseCredentials
 import com.intellij.database.dataSource.DatabaseAuthProvider
 import com.intellij.database.dataSource.DatabaseAuthProvider.ApplicabilityLevel
@@ -9,33 +10,43 @@ import com.intellij.database.dataSource.DatabaseConnectionConfig
 import com.intellij.database.dataSource.DatabaseConnectionInterceptor.ProtoConnection
 import com.intellij.database.dataSource.DatabaseConnectionPoint
 import com.intellij.database.dataSource.DatabaseCredentialsAuthProvider
+import com.intellij.database.dataSource.ui.AuthWidgetBuilder
 import com.intellij.database.dataSource.url.template.MutableParametersHolder
 import com.intellij.database.dataSource.url.template.ParametersHolder
 import com.intellij.openapi.project.Project
-import com.intellij.ui.components.JBLabel
-import com.intellij.ui.components.JBPasswordField
-import com.intellij.ui.components.JBTextField
-import com.intellij.uiDesigner.core.GridLayoutManager
-import kotlinx.coroutines.CoroutineName
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.future.future
-import java.util.concurrent.CompletionStage
-import javax.swing.JPanel
+import kotlinx.coroutines.withContext
 
 @Suppress("UnstableApiUsage")
-class VaultAuth : DatabaseAuthProvider, CoroutineScope {
+class VaultAuth : DatabaseAuthProvider {
     private val vault = Vault()
 
     companion object {
         private const val VAULT_ADDRESS = "vault.address"
         private const val VAULT_TOKEN = "vault.token"
         private const val VAULT_PATH = "vault.path"
+
+        private val noUrlHandler = object : AuthWidgetBuilder.UrlHandler<String> {
+            override fun fromUrl(holder: ParametersHolder): String = ""
+            override fun toUrl(value: String, holder: MutableParametersHolder) {}
+        }
+
+        private val noPasswordUrlHandler = object : AuthWidgetBuilder.UrlHandler<OneTimeString> {
+            override fun fromUrl(holder: ParametersHolder): OneTimeString = OneTimeString("")
+            override fun toUrl(value: OneTimeString, holder: MutableParametersHolder) {}
+        }
+
+        private val tokenSerializer = object : AuthWidgetBuilder.Serialiser<OneTimeString> {
+            override fun save(value: OneTimeString, config: DatabaseConnectionConfig, credentials: DatabaseCredentials?) {
+                config.dataSource.setAdditionalProperty(VAULT_TOKEN, value.toString())
+            }
+
+            override fun load(point: DatabaseConnectionPoint, credentials: DatabaseCredentials?): OneTimeString {
+                return OneTimeString(point.dataSource.getAdditionalProperty(VAULT_TOKEN) ?: "")
+            }
+        }
     }
 
-
-    override val coroutineContext = SupervisorJob() + Dispatchers.IO + CoroutineName("VaultAuth")
     override fun getId() = "vault"
 
     override fun getDisplayName() = VaultBundle.message("name")
@@ -45,15 +56,31 @@ class VaultAuth : DatabaseAuthProvider, CoroutineScope {
         level: ApplicabilityLevel
     ): ApplicabilityLevel.Result = ApplicabilityLevel.Result.APPLICABLE
 
-    @Deprecated("Use configureWidget() instead")
-    override fun createWidget(
+    override fun AuthWidgetBuilder.configureWidget(
         project: Project?,
         credentials: DatabaseCredentials,
         config: DatabaseConnectionConfig
-    ): DatabaseAuthProvider.AuthWidget = VaultWidget()
+    ) {
+        this
+            .addTextField(
+                { VaultBundle.message("addressLabel") },
+                AuthWidgetBuilder.additionalPropertySerializer(VAULT_ADDRESS),
+                noUrlHandler
+            )
+            .addPasswordField(
+                { VaultBundle.message("tokenLabel") },
+                tokenSerializer,
+                false,
+                noPasswordUrlHandler
+            )
+            .addTextField(
+                { VaultBundle.message("pathLabel") },
+                AuthWidgetBuilder.additionalPropertySerializer(VAULT_PATH),
+                noUrlHandler
+            )
+    }
 
-    @Deprecated("Use coroutines", replaceWith = ReplaceWith("interceptConnection"))
-    override fun intercept(proto: ProtoConnection, silent: Boolean): CompletionStage<ProtoConnection> {
+    override suspend fun interceptConnection(proto: ProtoConnection, silent: Boolean): Boolean {
         val mountPath = proto.connectionPoint.getAdditionalProperty(VAULT_PATH)
             ?: throw VaultAuthException(VaultBundle.message("invalidMountPath"))
         val addressPath = proto.connectionPoint.getAdditionalProperty(VAULT_ADDRESS)
@@ -61,81 +88,23 @@ class VaultAuth : DatabaseAuthProvider, CoroutineScope {
         val tokenPath = proto.connectionPoint.getAdditionalProperty(VAULT_TOKEN)
             ?: throw VaultAuthException(VaultBundle.message("invalidTokenPath"))
 
-        return future {
-            val json = try {
+        val json = try {
+            withContext(Dispatchers.IO) {
                 vault.readJson(mountPath, addressPath, tokenPath)
-            } catch (err: JsonProcessingException) {
-                throw VaultAuthException(VaultBundle.message("jsonError"), err)
             }
-
-            val username = json.path("data").path("username").asText()
-            val password = json.path("data").path("password").asText()
-
-            if (username.isEmpty() || password.isEmpty()) {
-                throw VaultAuthException(VaultBundle.message("invalidResponse"))
-            }
-
-            DatabaseCredentialsAuthProvider.applyCredentials(
-                proto,
-                Credentials(username, password),
-                true
-            )
-        }
-    }
-
-    @Suppress("TooManyFunctions", "EmptyFunctionBlock", "MagicNumber")
-    private class VaultWidget : DatabaseAuthProvider.AuthWidget {
-        private val addressField = JBTextField()
-        private val tokenField = JBPasswordField()
-        private val pathField = JBTextField()
-        private val panel = JPanel(GridLayoutManager(3, 6)).apply {
-            val addressLabel = JBLabel(VaultBundle.message("addressLabel"))
-            add(addressLabel, createLabelConstraints(0, 0, addressLabel.preferredSize.getWidth()))
-            add(addressField, createSimpleConstraints(0, 1, 3))
-
-            val tokenLabel = JBLabel(VaultBundle.message("tokenLabel"))
-            add(tokenLabel, createLabelConstraints(1, 0, tokenLabel.preferredSize.getWidth()))
-            add(tokenField, createSimpleConstraints(1, 1, 3))
-
-            val pathLabel = JBLabel(VaultBundle.message("pathLabel"))
-            add(pathLabel, createLabelConstraints(2, 0, pathLabel.preferredSize.getWidth()))
-            add(pathField, createSimpleConstraints(2, 1, 3))
+        } catch (err: JsonProcessingException) {
+            throw VaultAuthException(VaultBundle.message("jsonError"), err)
         }
 
-        override fun onChanged(r: Runnable) {}
+        val username = json.path("data").path("username").asText()
+        val password = json.path("data").path("password").asText()
 
-        override fun save(config: DatabaseConnectionConfig, copyCredentials: Boolean) {
-            with(config.dataSource) {
-                setAdditionalProperty(VAULT_PATH, pathField.text)
-                setAdditionalProperty(VAULT_ADDRESS, addressField.text)
-                setAdditionalProperty(VAULT_TOKEN, String(tokenField.password))
-            }
+        if (username.isEmpty() || password.isEmpty()) {
+            throw VaultAuthException(VaultBundle.message("invalidResponse"))
         }
 
-        override fun reset(point: DatabaseConnectionPoint, resetCredentials: Boolean) {
-            with(point.dataSource) {
-                pathField.text = getAdditionalProperty(VAULT_PATH) ?: ""
-                addressField.text = getAdditionalProperty(VAULT_ADDRESS) ?: ""
-                tokenField.text = getAdditionalProperty(VAULT_TOKEN) ?: ""
-            }
-        }
-
-        override fun updateUrl(model: MutableParametersHolder) {}
-
-        override fun updateFromUrl(holder: ParametersHolder) {}
-
-        override fun isPasswordChanged(): Boolean = false
-
-        override fun hidePassword() {}
-
-        override fun reloadCredentials() {}
-
-        override fun getComponent() = panel
-
-        override fun getPreferredFocusedComponent() = pathField
-
-        override fun forceSave() {}
-
+        DatabaseCredentialsAuthProvider.applyCredentials(proto, Credentials(username, password), true)
+        return true
     }
 
     internal class VaultAuthException(msg: String, cause: Throwable? = null) : RuntimeException(msg, cause)